ablation

+++ models/layers/sparsetopp_multihead_attention.py
@@ -81,9 +81,9 @@ class SparseTopPMultiheadAttention(nn.MultiheadAttention):
         self.attention_prune_rate=args.attention_prune_rate
 
 
-        self.q_act_mask=torch.randperm(self.embed_dim*self.embed_dim)[:(1-self.attention_prune_rate)*embed_dim*embed_dim]
-        self.k_act_mask=torch.randperm(self.kdim*self.embed_dim)[:(1-self.attention_prune_rate)*kdim*embed_dim]
-        self.v_act_mask=torch.randperm(self.vdim*self.embed_dim)[:(1-self.attention_prune_rate)*vdim*embed_dim]
+        self.q_act_mask=torch.randperm(self.embed_dim*self.embed_dim)[:int((1-self.attention_prune_rate)*embed_dim*embed_dim)]
+        self.k_act_mask=torch.randperm(self.kdim*self.embed_dim)[:(1-int(self.attention_prune_rate)*kdim*embed_dim)]
+        self.v_act_mask=torch.randperm(self.vdim*self.embed_dim)[:(1-int(self.attention_prune_rate)*vdim*embed_dim)]
 
         # Functionals
         self.q_scaling_product = nnq.FloatFunctional()
